+++ pkg/agent/storage/ci/keyvalue_store_test.go
@@ -1,6 +1,9 @@
 package storageci
 
 import (
+	"bytes"
+	"encoding/json"
+	"fmt"
 	"sync"
 	"testing"
 
@@ -152,4 +155,126 @@ func Test_Delete(t *testing.T) {
 			}
 		})
 	}
+}
+
+func Test_Updates(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		name    string
+		updates []map[string]string
+		want    []map[string]string
+	}{
+		{
+			name:    "empty",
+			updates: []map[string]string{{}, {}},
+			want:    []map[string]string{},
+		},
+		{
+			name:    "single",
+			updates: []map[string]string{{"one": "one"}, {"one": "new_one"}},
+			want: []map[string]string{
+				{
+					"key":   "one",
+					"value": "new_one",
+				},
+			},
+		},
+		{
+			name: "multiple",
+			updates: []map[string]string{
+				{
+					"one":   "one",
+					"two":   "two",
+					"three": "three",
+				},
+				{
+					"one":   "new_one",
+					"two":   "new_two",
+					"three": "new_three",
+				},
+			},
+			want: []map[string]string{
+				{
+					"key":   "one",
+					"value": "new_one",
+				},
+				{
+					"key":   "two",
+					"value": "new_two",
+				},
+				{
+					"key":   "three",
+					"value": "new_three",
+				},
+			},
+		},
+		{
+			name: "delete stale keys",
+			updates: []map[string]string{
+				{
+					"one":   "one",
+					"two":   "two",
+					"three": "three",
+					"four":  "four",
+					"five":  "five",
+					"six":   "six",
+				},
+				{
+					"four": "four",
+				},
+			},
+			want: []map[string]string{
+				{
+					"key":   "four",
+					"value": "four",
+				},
+			},
+		},
+	}
+	for _, tt := range tests {
+		tt := tt
+		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
+
+			for _, s := range getStores(t) {
+				for _, update := range tt.updates {
+					updateBytes, err := json.Marshal(update)
+					require.NoError(t, err)
+
+					s.Update(bytes.NewReader(updateBytes))
+				}
+
+				kvps, err := getKeyValueRows(s, tt.name)
+				require.NoError(t, err)
+
+				assert.ElementsMatch(t, tt.want, kvps)
+
+				for _, row := range kvps {
+					k := row["key"]
+					v := row["value"]
+
+					g, err := s.Get([]byte(k))
+					assert.NoError(t, err)
+					assert.Equal(t, []byte(v), g)
+				}
+			}
+		})
+	}
+}
+
+func getKeyValueRows(store types.KVStore, bucketName string) ([]map[string]string, error) {
+	results := make([]map[string]string, 0)
+
+	if err := store.ForEach(func(k, v []byte) error {
+		results = append(results, map[string]string{
+			"key":   string(k),
+			"value": string(v),
+		})
+		return nil
+	}); err != nil {
+		return nil, fmt.Errorf("fetching data: %w", err)
+	}
+
+	return results, nil
 }
